Fixed postgres connection issue.
Connections were timing out when the DB wasn't being used, and the
initial hit to the site after a time would result in a 500 error.
This is fixed by adding pool_pre_ping argument to the create_engine
call. This runs a SELECT 1 as the first statement prior to creating
a transaction and if that SELECT 1 fails, it will re-establish the
connection to the database before attempting to run the application
queries.